add notes to the ontology (#49)
This commit adds notes to classes and properties

NOTE: there is a chance that this commit has inadvertently reintroduced a deleted class/property. Please yell at me if that's the case and I can figure it out and fix it. Sorry about this, git is not good at merging such changes!

Co-authored-by: russell-d-e <62702525+russell-d-e@users.noreply.github.com>