adds colored output to rack.py, using colorama
This commit proposes a solution for adding colors to the output of
rack.py.

By overridding the logging mechanism, the code needs not change, except
the file-local "logger" should be used rather than the global root
logger "logging".

If this concept is accepted, there is more work to be done in replacing
some instances of "print" with proper logging commands.