Use specific property targets instead of base classes.
Using base classes as property targets introduces ambiguity and
removes the ability to provide any reasoning or Jena inference
regarding the relationships in the ontology (a specific class can
always be demoted to a base class, but the reverse is not true).  In
addition, it means that the actual structure is dictated by the user
data entered into the system rather than the ontology.

Now that we have identified the mechanisms by which we intend to
control the core ontology and support expansion and extension in
response to needs from users, we can remove this underlying ambiguity
from our specification.