Add query nodegroups for software structure
`Compilation Inputs.json`:

This query shows the inputs used in the compilation of the various executables,
shared libraries, etc. in the Turnstile example.

`Control Flow From Function.json`:

This query takes a runtime-specified function name, and shows the functions that
are called by the specified function. Try it out by entering "main" as the
function name.

`Files of a Given Format.json`:

This query takes a runtime-specified format name, and shows all the files of
that format involved in the Turnstile project. Try it out with different values
like `FORMAT-ElfFile`, `FORMAT-ZipFile`, and `FORMAT-CSourceFile`.