ada: gprbuild needs a configured xmlada source
It turns out that gprbuild needs to have access to a source tree for
xmlada that contains some files that are only generated after
`configure` is run.

However, `configure` only succeeds when `gprbuild` exists.  So the
bootstrap procedure is even more convoluted:

- build xmlada-source (just the source, not configured)
- build gprbuild-bootstrap (with that xmlada-source)
- build xmlada-source-configured (using gprbuild-bootstrap)
- build gprbuild (with that xmlada-source-configured)

There is still a problem remaining, however this one seems to be about
how Ada packages are imported rather than about files being missing.