ada: unify gnat, attempt to build ASIS
This commit makes sure every package depends on gnat, so that the
version of gnat can be changed uniformly.

Still trying to build ASIS (asis-for-gnat.nix), but encountering issues
with gnat_util, which seems to want to write to its build directory
post-install.  Trying to come up with workarounds.

This also includes a gcc-build-output package which compiles gcc, and
then saves its build directory snapshot in a build/ subdirectory of its
output, because gnat_util wants to find a file in there that is the
instantiation of an Ada template as far as I can tell.  Time-consuming
to build, but I have not yet found the incantation to just instantiate
the template and be done.