ada: connect static call graph to rdf outputter
As a side change, I stopped tracking the enclosing namespace, since
libadalang seems to also be doing it for us (as witnessed when
requesting the canonical fully-qualified name for a node).

The visitor still creates local visitors for recursion because we
locally override the value of callable_being_defined.  We now also
pass the filename, shared among all local instances of the visitor.

Each visitor creates its own node/edge set, and the caller merges them
up, this could be made simpler by just sharing the nodes and edges
dictionaries.