Refine what CI workflows do
In rack-box-continuous.yml, push rack-box image to Docker Hub only on
merges or commits to master, not when checking pull requests.

In rack-continuous.yml, lint files on pull requests as well as pushes
and run on ubuntu-20.04 like rest of workflows.

In README-Docker-Hub.md, explain Docker resource settings don't need
to be changed if they aren't visible in Docker settings anyway.