Updates based on Pull request Comments
added/harmonized all the alias for the sadl files.

Updated the description for TOOL to include the possiblity of a
non-software tool.

Added identifiers that were missing on some instances.

Aligned the Use of the Turnstile SYSTEM to always be turnstile, not
turnstiles, and to use the same project specific subclasses. for both
instance definition in planning data and in the system design.