ada: simplify treatment of files vs. bodies
Quite a big refactoring to make it easier to keep track of what is a
file, what is a function/procedure, and what keys, names, and URIs we
use for each case.

Also made is so that the local visitors share the nodes and edges with
their parent visitor, so that there is no longer need to merge upwards
when a visitor is done.

Finally, I've also made sure that we can process projects that do not
have a gpr project file.  In such cases, at the moment, all files
necessary for name resolution must be passed.  We may want to relax this
restriction and authorize dangling references to get partial results.