assist: remove generated Prolog files, update doc
Since the files are generated from the ontology, it is not necessary to
check them in, as they get out of date quickly.  Instead, the user can
read the updated README.md for information on how to regenerate the
files, and we will try and provide the generated files in the VMs.