ada: continue analysis upon unresolved references
The initial design was failing upon unresolved references, but
realistically, they are bound to happen.

We now continue the analysis, though we emit visible warnings, so as to
not give a false sense of completeness to the user when the analysis is
possibly missing details.

We will likely need to revisit the decision of what to do when we cannot
resolve all references in a project.