Can't download Fuseki 3.16.0 anymore and cache expired
Our cache of fuseki 3.16.0 expired even though we've been using it in
every CI build and we can't download fuseki 3.16.0 from the mirror
site since it's been replaced by fuseki 3.17.0.  Even though we could
download fuseki from <http://archive.apache.org/dist/jena/binaries/>
which keeps the old binaries around indefinitely, I think the download
link breaking is useful information that it's time to upgrade fuseki.

However, add -f (--fail) to the curl command so we'll notice
immediately when the download link breaks (not later on in the CI
workflow when we try to unpack the tarball).  Also stop caching the
fuseki tarball so we'll be compelled to upgrade fuseki more quickly
(otherwise we won't even notice a fuseki upgrade until weeks or months
later).  Switch from the mirror site to downloads.apache.org since
mirrors won't be as reliable as the original download site.