ada: emit warnings but continue on unresolved nodes
The previous behavior was to raise an exception when encountering a node
that could not be resolved to its definition, since this could lead to
issues of accidental captures.

Instead of giving up so dramatically, this commit adds visible warnings
in the logging output, and outputs such nodes with a mangled identifier
prefixed by '__RACK__UNRESOLVED__' so as to raise suspicion if such data
makes it into the database.

We will likely revisit this design decision once we have a better idea
of the use cases.