[assist] add parentheses to awk getline command input for mawk.
The original awk and gawk variation process the file fine, but the
mawk variation fails with strange errors on the original form:

```
/bin/sh: 1: dist.tar.gz: not found
/bin/sh: 1: /usr/share/locale/locale.alias: Permission denied
/bin/sh: 1: conf.json: not found
/bin/sh: 1: /etc/nsswitch.conf: Permission denied
/bin/sh: 1: /etc/ld.so.cache: Permission denied
/bin/sh: 1: /lib/x86_64-linux-gnu/libnss_files.so.2: Permission denied
/bin/sh: 1: /etc/passwd: Permission denied
/bin/sh: 1: /etc/group: Permission denied
/bin/sh: 1: libhw.so: not found
/bin/sh: 1: counter.exe: not found
```

It turns out this is due to a hyper-strict interpretation of the POSIX
awk spec as noted in
https://github.com/ThomasDickey/original-mawk/issues/35 which is
designated as a "won't fix" for mawk.  Adding the parentheses as in
this commit works around the mawk problem and all three variations now
work.