add experimental shell.nix for cli venv
This allows the automatic creation and registration of a virtualenv when
running `nix-shell` within the `cli` directory, mimicking the
environment you would get from creating the virtual environment in the
usual way.

Note that the current shell does not work well with software like
`lorri`, supposedly because the shell performs impure operations
(calling out to a server via `pip`).