cli: emit a warning guessing command to correct mistake
When one passes the wrong file to the command-line argument, we get a pretty
elaborate error message.  This adds a one-line warning suggesting a command
that would likely work given the fields in the input file.

That is, if the fields in the input file correspond to fields we expect in some
known command, this suggests the user to try that command instead.