ada: fix input to file provider
We allow passing a list of files to the file provider via a file listing
such files.  However, it is practical to have said file only contain
relative paths to the files it lists (relative to its own location).  As
such, we want to have the files be resolved w.r.t. the location of the
configuration file.  This commits adds support for this.