ada: use ScrapingToolKit as output
This large refactoring allows us to use the ScrapingToolKit to output
information from the analysis.

In order to not completely scrape away the Turtle output, and allow us
to later make different decisions, I tried to have a layer between the
analysis computing the output and the backend outputting it.

Because we were manipulating a lot of strings, including strings to be
used as keys in Python, strings to be used as keys in the ontology data,
strings to be used as names, ..., I also added 'NewType's that prevent
us from getting these mixed up, at the price of a bit of wrapping or
casting in some places.

Currently the output to Turtle is preserved behind a 'DEBUG_TURTLE'
flag, but it may make sense to either drop it entirely, or put it behind
a command-line argument later.