Remove EXPOSE line from Docker image
The EXPOSE line really doesn't do anything except document which ports
our rack-box image's services listen to.  We already document the
exact form of the "docker run" command that is needed to start running
a rack-box image successfully, so the EXPOSE line is just another
place that may become out of date later.