assist: write_ontology now lists all properties (#431)
There was a bug/limitation where write_ontology would only witness RACK
properties when some of the data was using the property.  As a result, we only
had a subset of all ontology properties.

This makes it so that all data and object properties in the ontology are now
listed.

Also fixes a bug where property names were lowercased in the export list but
not in their declaration.