assist: write_ontology now exports RDF assertions (#432)
assist: write_ontology now exports RDF assertions

The previous write_ontology exported what I refer to as RDF queries for each
class and property, that is, a Prolog predicate that allows one to retrieve
instances of a class of property that are registered within the database.

This adds a notion of RDF assertions for each class and property, allowing one
to register new data in the database.  As usual, this was already doable via
the ASSIST tooling, this commit just adds convenience operators for
RACK-specific data.