data_shape: use setup.cfg (#521)
I had trouble setting up the environment with the given setup.py.  I went to
read the documentation for setuptools, found out they recommended using
setup.cfg nowadays, switched the configuration for testing purposes, and it
worked.

It's very similar but makes the configuration a proper data file rather than
some constants in a piece of code.  If everyone is happy switching to
that, here is how we could go about it.