Generate and ingest dynamically-derived data for Turnstile.
The original Load-TurnstileData.sh script only ingests the static data
obtained from the various OWL/SADL files for Turnstile.  This updates
the script to also attempt to perform the dynamic build of the
Turnstile code with ASSIST process observation and subsequent data
upload.  This helps ensure the Turnstile example is complete for user
inspection.