migration: fix migration for SubsumeProperty
It turns out the migration steps to take for `SubsumeProperty` are a mix of the
ones for `RenameProperty` and the ones for `DeleteProperty`.  This makes sense
since, when a property is subsumed, it is both removed from the ontology, but
its references must be redirected just as if it had been renamed.