Replace unmaintained action
build-virtual-machine.yml: Call softprops/action-gh-release which is
the recommended replacement for actions/upload-release-asset which is
no longer maintained by the GitHub engineers.  Also split the virtual
machine first and upload exactly the same files using globbing to
either workflow artifacts or release assets to allow the split size
(-b 1500m) and corresponding number of split files to be tweaked
without having to change anything else in the steps.