Fix Build Virtual Machine Workflow
build-virtual-machine.yml: Good thing we made CI always split the
virtual machine since there was a typo in a filename that wouldn't
have been caught until it was time to make a release.  Fix mistake and
also add a comment warning that softprops/action-gh-release has many
issues and PRs filed against it, so we may need to replace it with "gh
release upload" CLI if action doesn't work when publishing a release.