generate auto-generated files during setup
STK is hard to use, because it expects users to have packages around
that are not listed in its requirements, to have a different version of
semtk-python3 than listed in its requirements, and it expects them to
run Generate-STK.py to get necessary files.

As a result, using STK as a dependency, as is done in the ada tooling,
is painful.

This commit updates dependencies to be more faithful, and attempts to
make the package install itself fully, running GenerateSTK.py as part of
its install process.

Note that I removed the dash from the name of the file as Python does
not particularly likes dashes in filenames.

While this requires having RACK set up prior to installing this
dependency, given that this package is somewhat unusable without those,
I think it's a price we ought to be willing to pay.